+++ test/integration/browser-rollup-esm/test.js
@@ -1,0 +1,5 @@
+describe('PROJ', () => {
+  it('PROJ quickstart', (done) => {
+    window.testDone = done;
+  });
+});