+++ wasm/index-node.mjs
@@ -3,9 +3,9 @@
  * import qPROJ from 'proj.js';
  * const PROJ = await qPROJ;
  * 
- * This is the browser specific WASM bundle
+ * This is the Node.js-specific WASM bundle
  */
-import bindings from '../lib/binding/emscripten-wasm32/proj.mjs';
+import bindings from '../lib/binding/emscripten-wasm32/proj_node.mjs';
 import emnapi from './emnapi.mjs';
 
 const result = emnapi(bindings);